requirements.txt
tensorflow
numpy
matplotlib
opencv-python
Pillow
streamlit
transformers
pytorch
langchain
langchain-huggingface
dotenv
requests

